+++ src/Datasets/Generators/Circle.php
@@ -27,7 +27,7 @@ class Circle implements Generator
      *
      * @var Vector
      */
-    protected \Tensor\Vector $center;
+    protected Vector $center;
 
     /**
      * The scaling factor of the circle.